Move hook definition to inline header
